requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.7
django-import-export==2.6.1
et-xmlfile==1.1.0
idna==3.3
lxml==4.6.3
MarkupPy==1.14
multitasking==0.0.9
numpy==1.21.2
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.3.4
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tablib==3.0.0
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
yfinance==0.1.63
reportlab
